phi_net: fix all six findings from Gemini independent audit, plus one it missed
A good audit. Six findings, all real. Four accepted at stated severity, two
accepted with severity corrected downward after I measured them.

The one that mattered: run_kaggle.py applied the FULL gate set, including
F1 > 0.70, to the B1 rung. B1 is diagnostic -- 100k rows, 15 epochs, answering
does Phi learn anything at all -- while F1 is judged at B2 on the test split. A
B1 of 0.66, a good result at that scale and exactly the case that should
proceed, would have aborted the Kaggle session before B2 ever started. My bug,
and a design error rather than a slip: I applied a falsification gate to a
diagnostic rung.

Fixed by extracting the decision into a pure function, b1_verdict(b1_auc,
material_auc), testable without a GPU or a 40-epoch run. Only two things now stop
the ladder: F0 failing, meaning the data is separable on piece counts so nothing
trained on it means anything; or Phi failing to beat the material baseline,
which is a representation result more epochs cannot fix.

Guarded by backend/tests/test_phi_net_gate.py -- seven tests, no GPU, no dataset
-- and MUTATION-CHECKED: reintroducing the original bug (if b1_auc <= 0.70) turns
test_b1_proceeds_when_signal_is_present_but_below_f1 red; restoring gives 12
passed. The same file pins roc_auc against a brute-force pairwise reference with
and without ties.

Also fixed: sys.path repair missing from evaluate.py; predict() defaulting
use_amp=True so --no-amp never reached validation or test inference; GradScaler
not following use_amp; block-buffered stdout that makes a healthy run look hung
for ninety epochs (flush=True in code, -u documented).

Two severities corrected by measurement rather than accepted. The float32
rank-sum concern is real in principle but assumed naive summation; torch sums
pairwise, and the measured error is 4e-08 to 8e-08 across all real sizes, five
decimal places below anything gates at 0.03 and 0.70 could notice. And a live
GradScaler with autocast off is not harmful: step() unscales before stepping and
scale-then-unscale by a power of two is exact in fp32, measured at 0.0 gradient
difference. Both fixed anyway, since both are free.

What the audit missed: run_kaggle.py had no --no-amp flag at all, so fixing the
threading inside train.py would still have left the documented mitigation
unreachable from the only entry point where it matters. Added.

Still unverified by anyone, and stated plainly in the response: the mixed
precision path has never executed on either side of this review. Gemini read it;
I could not run it. If Kaggle misbehaves, look there first -- and --no-amp now
genuinely bypasses all of it, which it did not before.

+++ phi_net/metrics.py
@@ -43,7 +43,13 @@ def roc_auc(y_true: torch.Tensor, score: torch.Tensor) -> float:
     ranks_sorted = group_mean_rank.repeat_interleave(counts)
     ranks = torch.empty_like(s)
     ranks[order] = ranks_sorted
-    return float((ranks[y == 1].sum() - n_pos * (n_pos + 1) / 2) / (n_pos * n_neg))
+    # Accumulate in float64. Rank sums reach ~1e10 at full-dataset sizes, past
+    # float32's exact-integer range of 2**24. Measured error was only ~5e-8
+    # because torch sums pairwise rather than sequentially -- irrelevant against
+    # gates at 0.03 and 0.70 -- but float64 costs nothing and removes the
+    # question.
+    return float((ranks[y == 1].double().sum() - n_pos * (n_pos + 1) / 2)
+                 / (n_pos * n_neg))
 
 
 def logistic_auc(x_train: torch.Tensor, y_train: torch.Tensor,
